+++ src/app/reducers/selectors.ts
@@ -6,4 +6,4 @@ export const getHistoryItems = (state: RootState): HistoryItem[] =>
   state.user.historyItems;
 
 export const getFirstLetterEmail = (state: RootState): string =>
-  state.user.email.slice(0, 1).toUpperCase();
+  state.user.email[0].toUpperCase();
